+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tutorlolv2"
-version = "0.4.4"
+version = "0.4.5"
 edition = "2024"
 authors = ["Luiz Gustavo Santana Dias Gomes"]
 repository = "https://github.com/LuizGomes56/tutorlolv2"
@@ -16,6 +16,7 @@ incremental = false
 
 [dev-dependencies]
 tutorlolv2_exports = { path = "tutorlolv2_exports" }
+tutorlolv2_build = { path = "tutorlolv2_build" }
 
 [dependencies]
 tutorlolv2_math = { path = "tutorlolv2_math" }
