+++ tutorlolv2_build/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tutorlolv2_build"
-version = "0.1.0"
+version = "0.1.1"
 edition = "2024"
 
 [profile.release]
@@ -14,8 +14,8 @@ incremental = false
 
 [dependencies]
 tutorlolv2_fmt = { path = "../tutorlolv2_fmt" }
-tokio = { version = "1.47", features = ["full"] }
 tutorlolv2_types = { path = "../tutorlolv2_types" }
+tokio = { version = "1.47", features = ["full"] }
 serde_json = "1.0.143"
 serde = { version = "1.0", features = ["derive"] }
 regex = "1.11.2"
