+++ tutorlolv2_build/src/scripts/export_items.rs
@@ -136,7 +136,7 @@ pub fn export_items() -> Vec<(u32, ItemDetails)> {
 
             let constdecl = format!(
                 r#"pub static {}:CachedItem=CachedItem {{gold:{},prettified_stats:&[{}],damage_type:{},attributes:Attrs::{},ranged:{},melee:{},stats:CachedItemStats{{{}}},}};"#,
-                format!("{}_{}", item.name.to_screaming_snake_case(), item_id),
+                format_args!("{}_{}", item.name.to_screaming_snake_case(), item_id),
                 item.gold,
                 prettified_stats,
                 if item.damage_type.is_some() {
