+++ tutorlolv2_dev/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tutorlolv2_dev"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2024"
 
 [dependencies]
@@ -12,7 +12,7 @@ once_cell = "1.21.3"
 scraper = "0.24.0"
 regex = "1.11.2"
 paste = "1.0.15"
-tokio = { version = "1.47", default-features = false }
+tokio = { version = "1.47", features = ["full"] }
 reqwest = { version = "0.12.1", features = [
     "json",
     "rustls-tls",
